==> 4_model_build.ipynb <==
#!pip3 install --upgrade sklearn 
#!pip3 uninstall xgboost


#The various imports
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os

import pandas as pd

import xgboost as xgb

cancelled_flights = pd.read_csv("data/all_flight_data_spark.csv")

cancelled_flights = cancelled_flights.dropna()

from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.preprocessing import OneHotEncoder, StandardScaler, LabelEncoder
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegressionCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import TransformerMixin
from sklearn.preprocessing import LabelEncoder, OrdinalEncoder
from sklearn.compose import ColumnTransformer
#from xgboost import XGBClassifier

X = cancelled_flights[['OP_CARRIER','OP_CARRIER_FL_NUM','ORIGIN','DEST','CRS_DEP_TIME','CRS_ELAPSED_TIME','DISTANCE','WEEK','HOUR']]

y = cancelled_flights[['CANCELLED']]

X.dtypes

# X['uniquecarrier'] = pd.Categorical(X['uniquecarrier'])
# X['flightnum'] = pd.Categorical(X['flightnum'])
# X['origin'] = pd.Categorical(X['origin'])
# X['dest'] = pd.Categorical(X['dest'])

#X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

categorical_cols = ['OP_CARRIER','OP_CARRIER_FL_NUM','ORIGIN','DEST']

ct = ColumnTransformer(
    [('le', OneHotEncoder(), categorical_cols)],
    remainder='passthrough'
)

X_trans = ct.fit_transform(X)

#X_trans = ct.transform(X) 


X_train, X_test, y_train, y_test = train_test_split(X_trans, y, random_state=42)

xgbclf = xgb.XGBClassifier() #LogisticRegressionCV(cv=cv, solver=solver, max_iter=max_iter)

pipe = Pipeline([('scaler', StandardScaler(with_mean=False)),
                 ('xgbclf', xgbclf)])

# The magical model.fit()
pipe.fit(X_train, y_train)

#xgbclf.fit(X_train, y_train)







train_score = pipe.score(X_train, y_train)
test_score = pipe.score(X_test, y_test)
print("train", train_score)
print("test", test_score)
#print(classification_report(y_test, pipe.predict(X_test)))
#data[labels.name + ' probability'] = pipe.predict_proba(X)[:, 1]

from joblib import dump, load

dump(pipe,'pipe.joblib')

dump(ct,'ct.joblib')
